# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: HackWerk HTTP API
  version: 0.4.0
  description: Interner Vertrag des serverseitig gerenderten HackWerk-Webdienstes.
servers:
  - url: http://localhost:18533
paths:
...
  /admin/users:
    get:
      operationId: listUsers
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Adminliste der Benutzer und Fahrerzuordnung. }
        "403": { description: Adminberechtigung fehlt. }
    post:
      operationId: createUser
      security: [{ sessionCookie: [] }]
      responses:
        "303": { description: Benutzer atomar angelegt und auditiert. }
        "403": { description: Adminberechtigung oder CSRF fehlt. }
...
  /availability:
    get:
      operationId: getOwnAvailability
      description: Mobile HTML-Verwaltung der eigenen Wochenregeln und Ausnahmen; das Zielprofil stammt ausschließlich aus der Session.
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Wochenplan, Ausnahmen und normalisierte 14-Tage-Vorschau in Europe/Vienna. }
        "403": { description: Kein zugeordnetes eigenes Fahrerprofil. }
  /admin/drivers:
    get:
      operationId: listDriverProfiles
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Fahrerprofile mit gemeinsamer Verfügbarkeitsübersicht der nächsten sieben Tage. }
        "403": { description: Adminberechtigung fehlt. }
    post:
      operationId: createDriverProfile
      security: [{ sessionCookie: [] }]
      responses:
        "303": { description: Eigenständiges, optional mit einem Login verknüpftes Fahrerprofil angelegt. }
        "403": { description: Adminberechtigung oder CSRF fehlt. }
        "422": { description: Profil oder Loginzuordnung ungültig. }
  /admin/resources:
    get:
      operationId: listResources
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Generische Maschinen- und Transportressourcen. }
        "403": { description: Adminberechtigung fehlt. }
    post:
      operationId: createResource
      security: [{ sessionCookie: [] }]
      responses:
        "303": { description: Ressource mit typisierten Kapazitätsdaten angelegt. }
        "403": { description: Adminberechtigung oder CSRF fehlt. }
        "422": { description: Typ, Name oder Kapazität ungültig. }
  /api/v1/me/availability:
    get:
      operationId: resolveOwnAvailability
      description: Liefert UTC-Intervalle für das Fahrerprofil der Session. Interne Notizen werden nie in das Overlay serialisiert.
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/AvailabilityFrom"
        - $ref: "#/components/parameters/AvailabilityTo"
      responses:
        "200":
          description: Normalisierte Intervalle samt minimaler Provenienz.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AvailabilityResponse" }
        "403": { description: Eigenes Fahrerprofil fehlt. }
        "422": { description: Zeitraum ungültig oder länger als 90 Tage. }
  /api/v1/drivers/{driverID}/availability:
    parameters:
      - $ref: "#/components/parameters/DriverID"
    get:
      operationId: resolveDriverAvailability
      description: Admin-Overlay eines Fahrerprofils; Fahrer können diesen Fremdpfad nicht verwenden.
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/AvailabilityFrom"
        - $ref: "#/components/parameters/AvailabilityTo"
      responses:
        "200":
          description: Normalisierte Intervalle samt minimaler Provenienz.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AvailabilityResponse" }
        "403": { description: Adminberechtigung fehlt. }
        "422": { description: Zeitraum ungültig oder länger als 90 Tage. }
...
components:
  parameters:
...
    DriverID:
      name: driverID
      in: path
      required: true
      schema: { type: string, format: uuid }
    AvailabilityFrom:
      name: from
      in: query
      required: true
      schema: { type: string, format: date-time }
    AvailabilityTo:
      name: to
      in: query
      required: true
      schema: { type: string, format: date-time }
...
  schemas:
...
    AvailabilityInterval:
      type: object
      additionalProperties: false
      required: [starts_at, ends_at, status, source, reason]
      properties:
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        status: { type: string, enum: [available, limited, unavailable] }
        source: { type: string, enum: [none, rule, exception, inactive_driver] }
        source_type: { type: string }
        reason: { type: string, description: Minimierter deutscher Grund ohne interne Bemerkung. }
    AvailabilityResponse:
      type: object
      additionalProperties: false
      required: [driver_id, timezone, intervals]
      properties:
        driver_id: { type: string, format: uuid }
        timezone: { type: string, const: Europe/Vienna }
        intervals:
          type: array
          items: { $ref: "#/components/schemas/AvailabilityInterval" }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hackplan_session